Incosistent license information
The POD claimed the license was "same as perl", but really it's GNU
APL. The latter is more permissive, so I don't think that will cause
any problems. In case you/somebody really want an older version to be
licensed under GPL-1 i don't really give a shit, feel free.