gh: utility to manage github repos
Having a special "github structure" in the filesystem, consisting
of $BASE_PATH/$AUTHOR/$REPO, this tool makes it easy to clone repos.
It automatically extracts the author and creates the author directory
if it doesn't exist, and more.

This is an initial commit, the tool is not working, but I use this
as a "test case" for the test framework.