gh-clone: github clone script
This script clones a git repository from github, in a path decided
based on the github account name. The base directory is specified
in the $REPO_BASE environment variable, which defaults to
$HOME/src/github.

By default, the script will clone using ssh, but you can override
that by setting the $GH_CLONE_URL_PREFIX, e.g.:
   GH_CLONE_URL_PREFIX=git://github.com/
Or...
   GH_CLONE_URL_PREFIX=https://github.com/
This same mechanism can also be used for github enterprise installs.

Final note: This script is designed so that it can be sourced, in
which case you will change directory to the repository when done.
To avoid the awkard . gh-clone syntax, you can alias this in your
~/.${SHELL}rc. For instance I do: alias gh='. gh-clone'

(I would avoid the temption of calling it ghc, since that's a haskell
compiler. Just fyi :).)