uri-encode: Add script to uri-encode from stdin
Be aware that we don't strip anything, so if you send in a trailing
newline, that will be encoded as well. If that's not what you want, use
`echo -n "$foo"` or `printf '%s' "$foo"`.

  echo -n 'foo' | uri-encode

You can create a symlink called uri-decode and point it to this script.
This will make it do the reverse, and decode uri encoded strings.

  echo -n '%30' | uri-decode