"wc-: Command not found" no more
If for some reason, any reason at all, you've made a habit of writing
`wc- l` instead of `wc -l`, this is the enterprise grade software
solution for you. Not only will it exec() the right program for you, it
will even interpret the first argument as flags.

We even provide a community edition for you freeloaders out there:
`alias wc-=wc`

                                    Community   Enterprise
 Remove "command not found" errors:    yes         yes
 On-premise                            yes         yes
 Interpret first argument as a flag    NO!         YES!

Please contact us for quotes on Enterprise licensing. No cloud version
as of yet. (How would that even look?  Are you ok with sending all of
your commands to some weird EC2 host?)

This tool may or may not be tightly coupled to wc, with wc being
hardcoded most notably (and possibly only) in the filename. Porting this
for use with other tools may not be feasible. I will not accept patches
implementing support for other tools.

This tool is licensed under the same terms as AT&T's implementation of
true, with sha1:da39a3ee5e6b4b0d3255bfef95601890afd80709.