hashsums-to-tree: Take a hashsum list and build a tree
This tools is meant to visualize duplicated files in a directory tree.
It takes the output format generated by md5sum, sha1sum, etc. and
outputs a YAML dict, with every path appearing as a list item on a list
keyed to the content hashsum. Complicated! But not really:

input:

> 6ffd97154dba59c71f4059dbcccff151  ./foo
> 98ca96b92091e8887455c61d3ba764f1  ./bar
> 385cda1cb77b6f1fc6b33cd82bb2de24  ./baz
> 385cda1cb77b6f1fc6b33cd82bb2de24  ./qux
>  385cda1cb77b6f1fc6b33cd82bb2de24  ./quux

output:

> ---
> 6ffd97154dba59c71f4059dbcccff151:
>  - ./foo
> 98ca96b92091e8887455c61d3ba764f1:
>  - ./bar
> 385cda1cb77b6f1fc6b33cd82bb2de24:
> - ./baz
> - ./qux
> - ./quux