cg3100-log: reflect "event log" of cable modem to syslog
Netgear CG3100 has a html table with an "event log" instead of a syslog.
This hack scrapes it and puts it in the local syslog instead.

I immediately put this in the deprecated directory, as i no longer have
access to the device.

The reason why i wrote this script led me to receive a new router from
my ISP. It should be noted in Netgear's defense, that the router had no
issue with reliability. My ISP (tele2) like the cargo cult leader they
are, decided to see if the neighborhood wide outage could be resolved by
replacing my "old" router. It didn't work of course, but now I have a
noname CPE router instead. Still no syslog.