unicodize-box: Support interior | in box
Commit @^{/Fix issue with multiple boxes on line} broke support for
having interior | characters within the box. This change attempts to
reintroduce this support by adding code to verify that the same start
and end positions are used on current line as is used on the previous.
If this isn't true for either of the terminators, we try to extend it
by moving it to the right.