unicodize-box: Improve demo (and progress indicator)
Not all of the boxes will be rendered by the script, but it serves to
actually demo the scripts current features and limitations (and cool to
see it improve over time! Just loop over all the commits touching this
file and run it with the current revision on master of the script as
input. Should be fun!).