osse: Remove broken arg scan
If a flag was supplied as first argument, it would enter an infinite
loop. The idea was to scan each argument for a non-flag and use that as
the name of the flag, but this was broken in more than one way. If this
were to be supported, we would have to consider flags that take
parameters as well, and for that, we would have to parse the... Yeah...

Besides, the aesthetics of not starting with the name of the flag is
questionable.